refactor(e2e): move getFirstConsultationLink to navigation helpers
It is a page/DOM navigation helper, so it belongs alongside
gotoFinaliseThemesList in navigation.ts rather than in the API-focused
helpers.ts. Update all consumers' imports; drop the now-unused Page
import from helpers.ts and a stale deleteFixtureData import from
consultations.e2e.ts.
